requirements.txt
torch --index-url https://download.pytorch.org/whl/cu118
torchvision
pandas
ipykernel
from_root
tqdm
matplotlib
seaborn
ensure==1.0.2
pyYAML
types-PyYAML

-e .